# Speed up loading trusted setup