--- a/src/c_kzg_4844.c
+++ b/src/c_kzg_4844.c
@@ -1600,7 +1600,7 @@ static C_KZG_RET expand_root_of_unity(
  * @param[in]  max_scale          Log base 2 of the number of roots of unity to
  *                                be initialized
  */
-static C_KZG_RET new_roots_of_unity(
+static C_KZG_RET compute_roots_of_unity(
     fr_t *roots_of_unity_out, uint32_t max_scale
 ) {
     C_KZG_RET ret;
@@ -1725,8 +1725,8 @@ C_KZG_RET load_trusted_setup(
         blst_p2_from_affine(&out->g2_values[i], &g2_affine);
     }
 
-    /* Initialize the KZGSettings struct */
-    ret = new_roots_of_unity(out->roots_of_unity, max_scale);
+    /* Compute roots of unity and permute the G1 trusted setup */
+    ret = compute_roots_of_unity(out->roots_of_unity, max_scale);
     if (ret != C_KZG_OK) goto out_error;
     ret = bit_reversal_permutation(out->g1_values, sizeof(g1_t), n1);
     if (ret != C_KZG_OK) goto out_error;