… bindings/java/src/testFixtures/resources/trusted-setups/trusted_setup.txt
… bindings/java/src/testFixtures/resources/trusted-setups/trusted_setup_4.txt
… src/test_c_kzg_4844.c
… src/trusted_setup.txt
… src/trusted_setup_4.txt