@zilm13 2023-05-28
I don't understand this change from the crypto side, but is there a way to check that we have new `trusted_setup` and throw if an old one provided? When using an old one I still produce some commitments and proofs, but they are not correct and fails on verification. So it would be great to throw if trusted_setup doesn't follow new expectations. Is it possible?